File scan FULLY works now
Parses JSON, takes the ID from the first JSON response and puts it into the second API JSON query, then proceeds to retrieve info about the analysis by VirusTotal, relays that to the user & tells if the file is malicious. Also provides links to more in-depth analysis which requires the API key used in the browser itself.