Take file with IPs as input + misc. changes
Instead of reading IP as input we read a file with IPs as input now. 

You can now input a file with IP addresses separated by commas, and it'll read through them and print out the results.

Debug prints are now set as comments instead (but still work just fine!)