FILE OUTPUT WORKS!
File output works. It clears the file whenever a new session is launched. Currently working on adding a random number or a timestamp appended to the filename to allow for multiple files to co-exist. For now, just take the output file and put it elsewhere & you're good to go (it checks for the file and removes it if it exists in the path of the program upon launch). :-)