SHODAN TOOL!
This tool allows you to scan IPs on Shodan, as well as find IP's running a certain service (via telling it which service you're looking for)

This program also highly focuses on being user friendly & prints in an easily-readable format.

I won't rule out that files can be considered as input and output rather than console prints in a future iteration, if that's desired. :)