Filter by Kaspersky DONE
It now filters by Kaspersky (I've made the code in a way that allows to easily change that AV) - and it also formats the print to let us know the findings of the AV we've chosen. 

I've also included an example output file with the IP's in the example input file. :-)